pyproject.toml

[project]
name = "biomedical-dashboards"
version = "0.1.0"
description = ""
authors = [{ name = "Keegan", email = "keegan.smith@curtin.edu.au" }]
readme = "README.md"
dependencies = [
    "pyyaml>=6.0, <7",
    "jinja2>=3.0",
    "google-cloud-bigquery>=3.0, <4",
    "GitPython>=3.0, <4"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["biomed", "queries"]

[project.scripts]
biomed = "main:main"


